+++ src/navigation/root/RootNavigation.tsx
@@ -3,10 +3,10 @@ import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { enableFreeze } from 'react-native-screens';
 
 import GlobalStackHeader from '@navigation/_components/styledHeader/GlobalStackHeader';
+import LocationHeader from '@navigation/home/_components/locationHeader/LocationHeader';
 import HomeNavigation from '@navigation/home/HomeNavigation';
 import { ROOT_NAVIGATION_LABEL_LIST } from '@navigation/root/RootNavigation.const';
 import { SettingNavigation } from '@navigation/setting/SettingNavigation';
-import LocationHeader from '@screens/HomeScreen/_components/locationHeader/LocationHeader';
 import LocationScreen from '@screens/LocationScreen/LocationScreen';
 import ModalScreen from '@screens/ModalScreen/ModalScreen/ModalScreen';
 import { useSettingStore } from '@store/settingStore/useSettingStore';
